Crashes in loadView()
In a UITableViewController sub class, the view property from UIViewController
and the tableView property from the UITableViewController point to the same object.
I wanted to create a new View to put the tableView in, so that i could constrain the tableview
to be always below the status bar. .... I am not sure what the correct way to do this is....